aarch64: Simplify guest memory setup
This change simplifies the guest memory region and page table setup on
aarch64 by:

 - Using a single KVM memory region for the entire guest physical memory.
 - Using a 2-level paging scheme with 1GB pages for the guest.

This brings the code in line with what the x86_64 code does, and removes
the need to have aarch64-specific code in the ELF loader.

Guest memory protection is _de facto_ unchanged and behaves the same way as
on x86_64; writing to a read-only section of the guest kernel (eg.
*solo5_app_main) will cause an exit from ukvm_hv_vcpu_loop() with "KVM:
host/guest translation fault" due to the underlying mprotect() performed by
the ELF loader.

I will add a test case to demonstrate this behaviour in a future change.