tenders/hvt: Arm64: Link used pmd tables to pud entries
In previous patch, we just linked pmd0 table to pud[0]. This is
the assumption of we are using no more than 1GB ram. But for
Arm64, we limit the ram size to 4GB. If users allocate 4GB ram
for application, when they access ram address higher than 1GB,
they will get page fault. In this patch, we linked used pmd tables
to pud entries to cover this case.

Signed-off-by: Wei Chen <wei.chen@arm.com>